frontend ui changed all input tag style changed and password eye button added and sidebar menu button added

+++ Backend/src/application/useCases/user/user/updateUserData.ts
@@ -39,7 +39,7 @@ export default class UpdateUserData {
     }
 
     let fileUrl;
-    if (user.profilePicture) {
+    if (file && user.profilePicture) {
       await this.awsS3Storage.deleteFile(user.profilePicture);
     }
     if (file) {
